Create Fighter.java
According to OOP, there are many types of inheritance, but Java supports only some of them (Multi-level and Hierarchical). To read more about it, please read this. Instructions In this exercise you're playing a role-playing game where different types of fighters can combat each other. The game has different rules for each type of fighter. We are going to focus on two specific types: Wizards and Warriors. For a Warrior, these are the rules: A Warrior is never vulnerable. A Warrior deals 6 points of damage if the fighter they are attacking is not vulnerable. A Warrior deals 10 points of damage if the fighter they are attacking is vulnerable. For a Wizard, these are the rules: A Wizard can prepare a spell in advance. A Wizard is vulnerable unless they have prepared a spell in advance. A Wizard deals 12 points of damage if they prepared a spell in advance.